BUGFIX: Correction of types used to represent time

+++ src/base/ProcessHit.cpp
@@ -61,12 +61,11 @@ EventBuffer<Hit> * ProcessHit::handleEvents (EventBuffer<RawHit> *inBuffer)
 			SystemConfig::TacConfig &ce = cc.tac_E[in.tacID];
 			SystemConfig::QacConfig &cq = cc.qac_Q[in.tacID];
 			SystemConfig::EnergyConfig &cen = cc.eCal[in.tacID];
-	       
 			
 			out.time = in.time;
 			if(useTDC) {
 				float q_T1 = ( -ct.a1 + sqrtf((ct.a1 * ct.a1) - (4.0f * (ct.a0 - in.t1fine) * ct.a2))) / (2.0f * ct.a2) ;
-				out.time = in.time - q_T1 - ct.t0;
+				out.time = double(in.time) - q_T1 - ct.t0;
 				out.qT1 = q_T1;
 				if(ct.a1 == 0) eventFlags |= 0x2;
 			}
@@ -74,7 +73,7 @@ EventBuffer<Hit> * ProcessHit::handleEvents (EventBuffer<RawHit> *inBuffer)
 			out.timeEnd = in.timeEnd;
 			if(useTDC) {
 				float q_T2 = ( -ce.a1 + sqrtf((ce.a1 * ce.a1) - (4.0f * (ce.a0 - in.t2fine) * ce.a2))) / (2.0f * ce.a2) ;
-				out.timeEnd = in.timeEnd - q_T2 - ce.t0;
+				out.timeEnd = double(in.timeEnd) - q_T2 - ce.t0;
 				out.qT2 = q_T2;
 				if(ce.a1 == 0) eventFlags |= 0x2;
 			}
